Tile grid component. ref #683
* Physics is set to 2D by default
* Collision object component has support for tile grids (grid shapes)
* Collision object resource can either use a conves chape or tile grid as shape input
* Added tests

+++ physics/src/physics/physics_2d.cpp
@@ -294,7 +294,7 @@ namespace dmPhysics
 
     HCollisionShape2D NewGridShape2D(HHullSet2D hull_set,
                                      const Vectormath::Aos::Point3& position,
-                                     uint32_t cell_width, uint32_t cell_height,
+                                     float cell_width, float cell_height,
                                      uint32 row_count, uint32 column_count)
     {
         b2Vec2 p(position.getX(), position.getY());
